+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [6.8.0]
+- Managed Object Storage support
+
 ## [6.7.1]
 
 ### Changed
@@ -15,7 +18,6 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - network: `dhcp_routes` field to IP network for additional DHCP classless static routes to be delivered if the DHCP is enabled
 - network: `static_routes` field to router for defining static routes
-- Managed Object Storage support
 
 ## [6.6.0]
 
@@ -346,7 +348,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.7.1...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.8.0...HEAD
+[6.8.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.7.1...v6.8.0
 [6.7.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.7.0...v6.7.1
 [6.7.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.6.0...v6.7.0
 [6.6.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.5.0...v6.6.0